# openapi/ciai-2.0.yaml
openapi: 3.1.0
info:
  title: CIAI Device Integration Protocol
  version: 2.0.0-beta.1
  description: Vendor-neutral API exposed by a CIAI-compatible laboratory device driver.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:12345
    description: Local development
paths:
...
  /HeartBeat:
    get:
      operationId: getHeartBeat
      summary: Get driver and device health
      responses:
        "200":
          description: Heartbeat result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartBeatResult"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /EnterAndExit:
    post:
      operationId: enterAndExit
      summary: Load, unload or transfer labware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnterOrExitData"
      responses:
        "200":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishResult"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  responses:
    BadRequest:
      description: Invalid JSON or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
    MethodNotAllowed:
      description: Known endpoint with an invalid HTTP method
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
...
  schemas:
    ResultBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: [string, "null"]
    ErrorResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data: {}
...
    HeartBeatResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/HeartBeatInfo"
...
    FinishResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              oneOf:
                - $ref: "#/components/schemas/Finish"
                - type: "null"
...
    EnterOrExitData:
      type: object
      required: [enterOrExitName]
      properties:
        enterOrExitName:
          type: string
          minLength: 1
        enterOrExitValue: {}
      additionalProperties: true
...
    Finish:
      type: object
      required: [completion]
      properties:
        completion:
          type: string
          enum: [finish, error]
        errorMsg:
          type: [string, "null"]
        instructionId:
          type: [string, "null"]
        nestId:
          type: [string, "null"]
        resultOutput:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/ResultOutput"
      additionalProperties: true
    ResultOutput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        resultData: {}
      additionalProperties: true
    HeartBeatInfo:
      type: object
      required: [heartBeatStatus, heartBeatTime]
      properties:
        heartBeatStatus:
          type: integer
          minimum: 0
          maximum: 6
        heartBeatTime:
          type: string
          format: date-time
      additionalProperties: true
...